% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCpromoter-package.R
\docType{package}
\name{CSCpromoter-package}
\alias{CSCpromoter}
\alias{CSCpromoter-package}
\title{CSCpromoter: CSCpromoter: Extract the sequences of promoter regions.}
\description{
Extract the sequences of promoter regions from the genome fasta files.
}
\keyword{internal}
