% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCpromoters-package.R
\docType{package}
\name{CSCpromoters-package}
\alias{CSCpromoters}
\alias{CSCpromoters-package}
\title{CSCpromoters: CSCpromoters: Extract the sequences of promoter regions}
\description{
Extract the sequences of promoter regions from the genome fasta files.
}
\keyword{internal}
