% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_locus.R
\name{filter_locus.character}
\alias{filter_locus.character}
\title{Filter annotation table by loci names}
\usage{
filter_locus.character(.annotations, .keep, .locus_var = "locus_tag")
}
\arguments{
\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described \link{HERE}. \code{data.table}
computations are normally faster.}

\item{.keep}{A \code{character} vector of loci names (present in the
\code{.annotations} table).}

\item{.locus_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing loci names. This can be
altered to fit input annotation tables with different
column names (not recommended). Defaults to \code{"locus_tag"}
.}
}
\description{
Filter annotation table by loci names
}
