% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_locus.R
\name{filter_locus.numeric}
\alias{filter_locus.numeric}
\title{Filter annotation table by loci indexes}
\usage{
filter_locus.numeric(.annotations, .keep = NULL)
}
\arguments{
\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described (HERE). \code{data.table}
computations are normally faster.}

\item{.keep}{A \code{numeric} vector of indexes (of the \code{.annotations} table).}
}
\description{
Filter annotation table by loci indexes
}
