% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_locus.R
\name{filter_locus}
\alias{filter_locus}
\title{Filter annotation table by loci names or loci indexes}
\usage{
filter_locus(.annotations, .keep = NULL, .locus_var = "locus_tag")
}
\arguments{
\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described {HERE}. \code{data.table}
computations are normally faster.}

\item{.keep}{Either a \code{character} vector of loci names (present in the
\code{.annotations} table) or a \code{numeric} vector of indexes
(of the \code{.annotations} table).}

\item{.locus_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing loci names. This can be
altered to fit input annotation tables with different
column names (not recommended). Defaults to \code{"locus_tag"}
.}
}
\value{
An annotations table object with the same class as \code{.annotations}
}
\description{
Filter annotation table by loci names or loci indexes
}
