% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plotdata_locus.R
\name{get_plotdata_locus}
\alias{get_plotdata_locus}
\title{get_plotdata_locus}
\usage{
get_plotdata_locus(
  .promoter_sizes,
  .annotations,
  .locus_var = "locus_tag",
  .chr_var = "chr",
  .closest_var = "closest_locus",
  .strand_var = "strand",
  .start_var = "begin",
  .end_var = "end"
)
}
\arguments{
\item{.promoter_sizes}{{TODO}}

\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described {HERE}. \code{data.table}
computations are normally faster.}

\item{.locus_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing loci names. This can be
altered to fit input annotation tables with different
column names (not recommended). Defaults to \code{"locus_tag"}
.}

\item{.chr_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing chromosome names. This
can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"chr"}.}

\item{.closest_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing closest loci names. This
can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"locus_tag"}.}

\item{.strand_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci strand
direction. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"strand"}.}

\item{.start_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci start
coordinate. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"begin"}.}

\item{.end_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci end coordinate.
This can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"end"}.}
}
\description{
get_plotdata_locus
}
