% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_promoter_sequences.R
\name{get_promoter_sequences}
\alias{get_promoter_sequences}
\title{Extract promoter sequences from the final promoter sizes.}
\usage{
get_promoter_sequences(
  .distances,
  .folder = NULL,
  .txdb = NULL,
  .pb = NULL,
  .pb_format = pb_format,
  .locus_var = "locus_tag",
  .chr_var = "chr",
  .dist_var = "promoter_size",
  .downstream = 0
)
}
\arguments{
\item{.distances}{}

\item{.folder}{}

\item{.txdb}{}

\item{.pb}{}

\item{.pb_format}{}

\item{.locus_var}{}

\item{.chr_var}{}

\item{.dist_var}{}

\item{.downstream}{}
}
\value{
DNAStringSet
}
\description{
Extract promoter sequences from the final promoter sizes.
}
