% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_promoters.R
\name{get_promoters}
\alias{get_promoters}
\title{Wrap-up function to query promoter sequences}
\usage{
get_promoters(
  .annotations,
  .keep = NULL,
  .txdb = NULL,
  .FASTA_list = NULL,
  .min_size = 100,
  .max_size = 2000,
  .downstream = 0,
  .pb = NULL,
  .pb_format = ":what - [:bar] :percent (:spin)",
  .locus_var = "locus_tag",
  .closest_var = "closest_locus",
  .chr_var = "chr",
  .strand_var = "strand",
  .start_var = "begin",
  .end_var = "end",
  .dist_var = "dist",
  .size_var = "promoter_size",
  .distances = NULL,
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described {HERE}. \code{data.table}
computations are normally faster.}

\item{.keep}{Either a \code{character} vector of loci names (present in the
\code{.annotations} table) or a \code{numeric} vector of indexes
(of the \code{.annotations} table).}

\item{.txdb}{{TODO}}

\item{.FASTA_list}{A named \code{character} list. Each element is a full path and
filename to each chromosome. Element names indicate
chromosome names matching those found in the annotation
table.}

\item{.min_size}{An integer indicating the minimum promoter size. Any
\emph{loci} identified to have sequence length shorter than
\code{.min_size} will be ignored.}

\item{.max_size}{An integer indicating the maximum promoter size. Any
\emph{loci} identified to have sequence length longer than
\code{.max_size} will be set (trimmed) to equals \code{.max_size}.}

\item{.downstream}{An integer indicating the sequence length downstream the
start site to be included in the output promoter
sequence. Presumably, will be reduced from the upstream
sequence length but have never tested. Defaults to \code{0}.}

\item{.pb}{An optional progress bar object created via
\code{progress::progress_bar()}. Defaults to \code{NULL}, case in
which a progress bar will be generated internally.}

\item{.pb_format}{A \code{character} indicating the format of the progress bar.
A number of tokens can be used here, see them below. It
defaults to \code{":what - [:bar] :percent (:spin)"}, which
means that information on the locus being handled in
shown on the left (\verb{:what}), followed by the progress
bar is within brackets (\verb{[:bar]}), and ended by the
percentage and a spinning character on the right
(\verb{:percent (:spin)}).}

\item{.locus_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing loci names. This can be
altered to fit input annotation tables with different
column names (not recommended). Defaults to \code{"locus_tag"}
.}

\item{.closest_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing closest loci names. This
can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"locus_tag"}.}

\item{.chr_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing chromosome names. This
can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"chr"}.}

\item{.strand_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci strand
direction. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"strand"}.}

\item{.start_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci start
coordinate. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"begin"}.}

\item{.end_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci end coordinate.
This can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"end"}.}

\item{.dist_var}{A \code{character} indicating the name of column used to
output the distance from each locus to its closest
upstream locus. I would bet you don't have a reason to
change that. Defaults to \code{"dist"}.}

\item{.size_var}{A \code{character} indicating the name of the output column
table containing the final promoter lengths. I still
doubt you need to change that. Defaults to
\code{"promoter_size"}.}

\item{.distances}{{TODO}}

\item{.parallel}{A \code{logical}. If TRUE, apply computations in parallel.
Defauts to \code{FALSE}.}

\item{...}{Future-proof stuff.}
}
\value{
DNAStringSet
}
\description{
Wrap-up function to query promoter sequences
}
\section{Tokens}{


They can be used in the \code{format} argument when creating the
progress bar.
\describe{
  \item{:bar}{The progress bar itself.}
  \item{:current}{Current tick number.}
  \item{:total}{Total ticks.}
  \item{:elapsed}{Elapsed time in seconds.}
  \item{:elapsedfull}{Elapsed time in hh:mm:ss format.}
  \item{:eta}{Estimated completion time in seconds.}
  \item{:percent}{Completion percentage.}
  \item{:rate}{Download rate, bytes per second. See example below.}
  \item{:tick_rate}{Similar to \code{:rate}, but we don't assume that
     the units are bytes, we just print the raw number of ticks per
     second.}
  \item{:bytes}{Shows :current, formatted as bytes. Useful
     for downloads or file reads if you don't know the size of the
     file in advance. See example below.}
  \item{:spin}{Shows a spinner that updates even when progress is
     advanced by zero.}
}

Custom tokens are also supported, and you need to pass their
values to \code{progress_bar$tick()} or \code{progress_bar$update()},
in a named list. See example below.

}

