% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_txdb.R
\name{make_txdb}
\alias{make_txdb}
\title{make a TxDb object from transcript annotations available as a GFF3 or GTF file.}
\usage{
make_txdb(.filename = NULL, .data_source = NULL, .organism = NULL)
}
\value{
A TxDb object.
}
\description{
make a TxDb object from transcript annotations available as a GFF3 or GTF file.
}
