% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_promoter_maps.R
\name{plot_promoter_maps}
\alias{plot_promoter_maps}
\title{plot_promoter_maps}
\usage{
plot_promoter_maps(.promoter_sizes, .annotations)
}
\arguments{
\item{.promoter_sizes}{{TODO}}

\item{.annotations}{Preferably a \code{data.table} (but possibly a \code{data.frame})
of annotations as described {HERE}. \code{data.table}
computations are normally faster.}
}
\description{
plot_promoter_maps
}
