% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_distances.R
\name{trim_distances}
\alias{trim_distances}
\title{Trim the length of promoters based on the distance to the closest upstream gene and the user-defined minimum and maximum length.}
\usage{
trim_distances(.distances, .min_size = 100, .max_size = 2000)
}
\arguments{
\item{.distances}{}

\item{.min_size}{}

\item{.max_size}{}
}
\value{
A data frame with filtered genes that have promoter lengths longer than the user-defined minimum length, and a new column containing the length of the final promoter sizes.
}
\description{
Trim the length of promoters based on the distance to the closest upstream gene and the user-defined minimum and maximum length.
}
