% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Export promoter sequences in a fasta file.}
\usage{
write_fasta(.sequences, .output_file, ...)
}
\arguments{
\item{.sequences}{{TODO}}

\item{.output_file}{{TODO}}

\item{...}{Future-proof stuff.}
}
\value{
A fasta file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
