% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rda.R
\name{write_rda}
\alias{write_rda}
\title{Export extracted promoter sequence results in a R Data File.}
\usage{
write_rda(.sequences, .output_file, ...)
}
\arguments{
\item{.sequences}{{TODO}}

\item{.output_file}{{TODO}}

\item{...}{Future-proof stuff.}
}
\value{
A R Data File \code{rda}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
