% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_promoter_distances.R
\name{get_promoter_distances}
\alias{get_promoter_distances}
\title{Get the distance from the gene-of-interest transcription start site to the closest upstream gene.}
\usage{
get_promoter_distances(
  .annotations,
  .pb = NULL,
  .pb_format = pb_format,
  .locus_var = "locus_tag",
  .chr_var = "chr"
)
}
\arguments{
\item{.annotations}{}

\item{.pb}{}

\item{.pb_format}{}

\item{.locus_var}{}

\item{.chr_var}{}
}
\value{
A data frame with the closest upstream gene ID and distance from the gene-of-interest transcription start site to the closest upstream gene.
}
\description{
Get the distance from the gene-of-interest transcription start site to the closest upstream gene.
}
