% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txdb_annotation.R
\name{get_txdb_annotation}
\alias{get_txdb_annotation}
\title{Extract annotation table from TxDB object}
\usage{
get_txdb_annotation(
  .txdb,
  .locus_var = "locus_tag",
  .chr_var = "chr",
  .strand_var = "strand",
  .start_var = "begin",
  .end_var = "end"
)
}
\arguments{
\item{.txdb}{{TODO}}

\item{.locus_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing loci names. This can be
altered to fit input annotation tables with different
column names (not recommended). Defaults to \code{"locus_tag"}
.}

\item{.chr_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing chromosome names. This
can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"chr"}.}

\item{.strand_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci strand
direction. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"strand"}.}

\item{.start_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci start
coordinate. This can be altered to fit input annotation
tables with different column names (not recommended).
Defaults to \code{"begin"}.}

\item{.end_var}{A \code{character} indicating the name of the column in the
\code{.annotation} table containing each loci end coordinate.
This can be altered to fit input annotation tables with
different column names (not recommended). Defaults to
\code{"end"}.}
}
\description{
Extract annotation table from TxDB object
}
