% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Export promoter sequences in a fasta file.}
\usage{
write_fasta(.sequences, .output_file)
}
\arguments{
\item{.sequences}{}

\item{.output_file}{}
}
\value{
A fasta file.
}
\description{
Export promoter sequences in a fasta file.
}
