% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rda.R
\name{write_rda}
\alias{write_rda}
\title{Export extracted promoter sequence results in a R Data File.}
\usage{
write_rda(.sequences, .output_file)
}
\value{
A R Data File \code{rda}.
}
\description{
Export extracted promoter sequence results in a R Data File.
}
